pyproject.toml
[project]
name = "datalens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jordan Bickelhaupt", email = "jrbickelh@gmail.com" }
]
requires-python = ">=3.12.8"
dependencies = [
    "chromadb>=1.5.1",
    "deltalake>=1.4.2",
    "duckdb>=1.4.4",
    "great-expectations>=1.12.3",
    "lancedb>=0.29.2",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langsmith>=0.7.5",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
    "tabulate>=0.9.0",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
